refactor: UserHobbyTag 경로 위치 변경, 회원가입 시 hobbyTag 이름 기반으로 매핑하도록 수정
