+++ src/main/java/swyp/hobbi/swyphobbiback/hobbytag/domain/HobbyTag.java
@@ -2,7 +2,7 @@ package swyp.hobbi.swyphobbiback.hobbytag.domain;
 
 import jakarta.persistence.*;
 import lombok.*;
-import swyp.hobbi.swyphobbiback.user.domain.UserHobbyTag;
+import swyp.hobbi.swyphobbiback.userhobbytag.domain.UserHobbyTag;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -28,5 +28,6 @@ public class HobbyTag {
 //    private List<PostHobbyTag> postHobbyTags = new ArrayList<>();
 
     @OneToMany(mappedBy = "hobbyTag", cascade = CascadeType.ALL, orphanRemoval = true)
+    @Builder.Default
     private List<UserHobbyTag> userHobbyTags = new ArrayList<>();
 }
