+++ src/main/java/swyp/hobbi/swyphobbiback/hobbytag/repository/HobbyTagRepository.java
@@ -4,6 +4,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import swyp.hobbi.swyphobbiback.hobbytag.domain.HobbyTag;
 
+import java.util.List;
+
 @Repository
 public interface HobbyTagRepository extends JpaRepository<HobbyTag,Long> {
+    List<HobbyTag> findAllByHobbyTagNameIn(List<String> hobbyTagNames);
 }
