+++ src/main/java/swyp/hobbi/swyphobbiback/user/domain/User.java
@@ -5,6 +5,7 @@ import lombok.*;
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
+import swyp.hobbi.swyphobbiback.userhobbytag.domain.UserHobbyTag;
 
 import java.time.LocalDateTime;
 import java.util.ArrayList;
